% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemplar.R
\name{assertions}
\alias{assertions}
\title{Generate character vector of assertions for a given object}
\usage{
assertions(x, data_name, ...)
}
\arguments{
\item{x}{The object for which to generate assertions.}

\item{data_name}{Name of data object to validate, as it would be called in
the interior of the function. This defaults to "data", but it may make
sense to use other names for individual assertions, eg. "data$column".}

\item{...}{Arguments passed to class-specific functions.}
}
\description{
Generate character vector of assertions for a given object
}
