% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemplar.R
\name{exemplar}
\alias{exemplar}
\title{Create a validation function from an exemplar}
\usage{
exemplar(x, ...)
}
\arguments{
\item{x}{The object for which to create an exemplar}

\item{...}{Additional arguments used when building the assertions. Currently
unused.}
}
\description{
Create a validation function from an exemplar
}
